requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-storages==1.10.1
gunicorn==20.0.4
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.1
whitenoise==5.2.0
wrapt==1.12.1
